requirements-full.txt
anthropic
httpx
beautifulsoup4
playwright
click
sqlite-vec
sentence-transformers
datasette
lxml
tqdm
python-dotenv
